Change id back from usize to u64 so that it is consistent over 64 and 32 environments

+++ src/id_based_blocking/cloneable_id_based_writer.rs
@@ -22,7 +22,7 @@ impl CloneableIdBasedWriter {
         }
     }
 
-    pub fn switch(&self, pool_item_id: usize) {
+    pub fn switch(&self, pool_item_id: u64) {
         let writer = unsafe { &mut *self.writer.get() };
         writer.set_pool_item(pool_item_id);
     }
