+++ src/samples/randoms/randoms_api/mean_response.rs
@@ -5,7 +5,7 @@ use super::RandomsApi;
 /// The response from a request to calculate the mean
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct MeanResponse {
-    pub id: usize,
+    pub id: u64,
     pub mean: u128,
 }
