+++ src/thread_request_response/remove_pool_item_response.rs
@@ -4,16 +4,16 @@ use super::ThreadRequestResponse;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct RemovePoolItemResponse {
-    id: usize,
+    id: u64,
     item_existed: bool,
 }
 
 impl RemovePoolItemResponse {
-    pub fn new(id: usize, item_existed: bool) -> Self {
+    pub fn new(id: u64, item_existed: bool) -> Self {
         Self { id, item_existed }
     }
 
-    pub fn id(&self) -> usize {
+    pub fn id(&self) -> u64 {
         self.id
     }
